+++ src/component/dom/render.ts
@@ -1,17 +1,27 @@
-import { Render, RenderRequest } from '../.';
 import { isDev } from '../../config';
+import { TagItem } from '../../client/TagifyClient'
+
+const DEBUG_PREFIX_DOM_RENDER = '[domRender]';
+
+/**
+ * Request for rendering tags.
+ */
+export interface RenderRequest {
+    target: Element;
+    tags: TagItem[];
+}
+
+export type Render = (request: RenderRequest) => void;
 
 /**
  * Renders a list of tags inside given target DOM element.
  */
-const render: Render = (request: RenderRequest) => {
+export const domRender: Render = (request: RenderRequest) => {
 
     const { target, tags } = request;
 
     if (isDev()) {
-        console.log('[DOM render] request target type: ' + target.type);
-        console.log('[DOM render] request target id: ' + target.id);
-        console.log('[DOM render] request tags: ' + JSON.stringify(tags));
+        console.log(`${DEBUG_PREFIX_DOM_RENDER} tags: ${JSON.stringify(tags)}`);
     }
 
     if (!target || tags.length == 0) {
@@ -19,7 +29,7 @@ const render: Render = (request: RenderRequest) => {
     }
 
     if (isDev()) {
-        console.log('[DOM render] rendering tags...');
+        console.log(`${DEBUG_PREFIX_DOM_RENDER} rendering tags...`);
     }
 
     const ul = document.createElement("ul");
@@ -34,15 +44,13 @@ const render: Render = (request: RenderRequest) => {
         a.href = tag.source;
         a.innerText = tag.value || `tag ${i}`;
         a.className = 'tagifyLink';
-        
+
         let li = document.createElement("li");
         li.className = 'tagifyRow';
         li.appendChild(a);
-        
+
         ul.appendChild(li);
     });
 
     target.appendChild(ul);
 };
-
-export default render;