+++ src/component/index.ts
@@ -1,0 +1,3 @@
+import getTags from "./dom";
+
+export default getTags;