+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.3.0
+ - exposed default `#tagify` method as `#getTags`;
+ - added `#getTagsForAnchors` as a quality of life improvement ofr ease of use;
+ - narrowed down `lodash` dependency only to `lodash.noop`.
+
 ## 0.2.0
  - breaking change: added support for batch tag rendering, i.e. for multiple targets.